requirements.txt
python-nmap
manuf
